Working on the crossover
There is a problem yet that the vector of the points x is basically a binary number, and that will be taken into consideration soon. Furthermore, it is not known yet how many times the roulette wheel should spin. Finally, there will be work done in near future on the reasonable fittnes function